requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
python-multipart==0.0.9
resend==2.0.0
apscheduler==3.10.4
